fix meds integration

+++ platforms/android/app/src/main/assets/www/js/meds.js
@@ -5,15 +5,49 @@ function init_meds() {
     })
 
     $("body").delegate(".med-delete", "click", function(e) {
-        delete_med($(this).attr("med_id"))
+        swal({
+            title: "Remove medication?",
+            text: "Yes delete, remove this medication.",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true
+        }).then(willDelete => {
+            if (willDelete) {
+                delete_med($(this).attr("med_id"))
+            }
+        })
     })
 
+    if (localStorage.getItem("token")) {
+        refresh_meds()
+    }
+}
+
+function refresh_meds() {
     list_meds(function(meds) {
+        $("#med_list").empty()
         for(var med of meds) {
             add_med_html(med)
         }
     })
 }
+
+function handle_meds_click() {
+    if (localStorage.getItem("token")) {
+        hide_screens()
+        $("#my-meds-screen").show()
+    }
+    else {
+        swal({
+            title: "Error",
+            text: "You must first login to your profile",
+            icon: 'error',
+        })
+        $("#my-profile").click()
+    }
+}
+
+
 
 function add_med_html(med) {
     $("#med_list").append(`
@@ -52,7 +86,7 @@ function delete_med(med_id) {
           text: "Medication Removed",
           icon: "success",
         });
-
+        refresh_meds()
     }).fail(function (err) {
       alert("ERROR")
     });
@@ -84,6 +118,8 @@ function add_meds() {
           text: "Medication Added",
           icon: "success",
         });
+        refresh_meds()
+        $(".close").click()
 
     }).fail(function (err) {
       alert("ERROR")
